+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## v0.24.2 (2025-07-14)
+### :rocket: Enhancement
+* [#33](https://github.com/lblod/frontend-lpdc/pull/33) creator and last modifier filtering [LPDC-1421] ([@wolfderechter](https://github.com/wolfderechter))
+
+## v0.24.1 (2025-07-10)
+### :rocket: Enhancement
+* [#32](https://github.com/lblod/frontend-lpdc/pull/32) add yourEurope codelist ordering [LPDC-1423] ([@wolfderechter](https://github.com/wolfderechter))
+
+## v0.24.0 (2025-06-20)
+### :rocket: Enhancement
+* Re-release of the changes in v0.22.2 and v0.23.0 due to conflicting releases.
+
+## v0.23.0 (2025-06-18)
+### :rocket: Enhancement
+* [#27](https://github.com/lblod/frontend-lpdc/pull/27) Implement impersonation in ACM/IDM [LPDC-1029] ([@benjay10](https://github.com/benjay10))
+
+## v0.22.2 (2025-06-20)
+#### :bug: Bug Fix
+* [#30](https://github.com/lblod/frontend-lpdc/pull/30) refactor: validate authority levels on publish instead of on save [LPDC-1431] ([@wolfderechter](https://github.com/wolfderechter))
+
+## v0.22.1 (2025-05-28)
+#### :bug: Bug Fix
+* [#28](https://github.com/lblod/frontend-lpdc/pull/28) upgrade to 2.10.3 submission-form-helpers ([@wolfderechter](https://github.com/wolfderechter))
+
+## v0.22.0 (2025-05-16)
+#### :rocket: Enhancement
+* [#26](https://github.com/lblod/frontend-lpdc/pull/26) [LPDC-1359] add creator and lastmodifier user ([@wolfderechter](https://github.com/wolfderechter))
+* [#25](https://github.com/lblod/frontend-lpdc/pull/25) [LPDC-1411] upgrade to appuniversum v3 ([@wolfderechter](https://github.com/wolfderechter))
+* [#12](https://github.com/lblod/frontend-lpdc/pull/12) [LPDC-1412] Rich text editor update ([@Windvis](https://github.com/Windvis))
+
+## v0.21.1 (2025-04-15)
+#### :bug: Bug Fix
+* [#23](https://github.com/lblod/frontend-lpdc/pull/23) fix: show error messages before asking for confirmation to publish product instance ([@mirdono](https://github.com/mirdono))
+
+## v0.21.0 (2025-04-10)
+
+#### :rocket: Enhancement
+* [#20](https://github.com/lblod/frontend-lpdc/pull/20) feature: Add validation before saving ([@wolfderechter](https://github.com/wolfderechter))
+
+## v0.20.0 (2025-03-04)
+
+#### :rocket: Enhancement
+* [#18](https://github.com/lblod/frontend-lpdc/pull/18) feature: introduce fusies feature flag to show/hide fusies ([@wolfderechter](https://github.com/wolfderechter))
+* [#17](https://github.com/lblod/frontend-lpdc/pull/17) Refactor: use correct NamedNode constructor ([@mirdono](https://github.com/mirdono))
+
 ## v0.19.0 (2024-11-28)
 #### :rocket: Enhancement
 * [#16](https://github.com/lblod/frontend-lpdc/pull/16) [LPDC-1318] Enable address validation for product instances labelled for merger ([@mirdono](https://github.com/mirdono))